Silenced a shifting a negative signed value is undefined warning from clang.

+++ drbd/drbd_main.c
@@ -1926,7 +1926,7 @@ static void dcbp_set_start(struct p_compressed_bm *p, int set)
 static void dcbp_set_pad_bits(struct p_compressed_bm *p, int n)
 {
 	BUG_ON(n & ~0x7);
-	p->encoding = (p->encoding & (~0x7 << 4)) | (n << 4);
+	p->encoding = (p->encoding & (((uint8_t)~0x7) << 4)) | (n << 4);
 }
 
 /* For compat reasons, encode bitmap as if it was 4k per bit!
